# src/main/resources/openapi.yml
---
openapi: 3.0.3
info:
  title: jobs-service-public-api API
  version: 1.0.0-SNAPSHOT
paths:
  /jobs:
    post:
      tags:
        - Job Service Resource
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/JobDef'
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/JobDef'
  /jobs/find:
    get:
      tags:
        - Job Service Resource
      parameters:
        - name: businessKey
          in: query
          schema:
            default: ""
            type: string
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/JobDef'
  /jobs/{id}:
    get:
      tags:
        - Job Service Resource
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/JobDef'
components:
  schemas:
    CronScheduleDef:
      description: Schedule definition to execute a job on a cron based configuration
      type: object
      allOf:
        - $ref: '#/components/schemas/ScheduleDef'
      properties:
        expression:
          type: string
        timeZone:
          type: string
    HttpRecipientDef:
      description: Recipient definition to execute a job with a http request
      type: object
      allOf:
        - $ref: '#/components/schemas/RecipientDef'
      properties:
        url:
          type: string
        method:
          type: string
        headers:
          type: object
          additionalProperties:
            type: string
        queryParams:
          type: object
          additionalProperties:
            type: string
    JobDef:
      description: Defines a job that can be managed by the Jobs Service
      type: object
      properties:
        id:
          type: string
        businessKey:
          type: string
        state:
          $ref: '#/components/schemas/StateDef'
        schedule:
          $ref: '#/components/schemas/ScheduleDef'
        retry:
          $ref: '#/components/schemas/RetryDef'
        recipient:
          $ref: '#/components/schemas/RecipientDef'
    KafkaRecipientDef:
      description: Recipient definition to execute a job with a kafka message
      type: object
      allOf:
        - $ref: '#/components/schemas/RecipientDef'
      properties:
        bootstrapServers:
          type: string
        topicName:
          type: string
        headers:
          type: object
          additionalProperties:
            type: string
    RecipientDef:
      required:
        - type
      type: object
      properties:
        type:
          type: string
        payload:
          type: object
      discriminator:
        propertyName: type
        mapping:
          http: '#/components/schemas/HttpRecipientDef'
          kafka: '#/components/schemas/KafkaRecipientDef'
          sink: '#/components/schemas/SinkRecipientDef'
    RetryDef:
      description: The retry configuration establishes the number of times a failing
        Job execution must be retried before it’s considered as FAILED
      type: object
      properties:
        maxRetries:
          format: int32
          type: integer
        delay:
          format: int64
          type: integer
        delayUnit:
          $ref: '#/components/schemas/TemporalUnitDef'
        maxDuration:
          format: int64
          type: integer
        durationUnit:
          $ref: '#/components/schemas/TemporalUnitDef'
    ScheduleDef:
      required:
        - type
      type: object
      properties:
        type:
          type: string
      discriminator:
        propertyName: type
        mapping:
          cron: '#/components/schemas/CronScheduleDef'
          timer: '#/components/schemas/TimerScheduleDef'
    SinkRecipientDef:
      description: Recipient definition to execute a job with a cloud event on a knative
        sink
      type: object
      allOf:
        - $ref: '#/components/schemas/RecipientDef'
      properties:
        sinkUrl:
          type: string
        contentMode:
          type: string
    StateDef:
      description: Available states for a Job
      enum:
        - TBD1
        - TBD2
        - TBD3
      type: string
    TemporalUnitDef:
      enum:
        - MILLIS
        - SECONDS
        - MINUTES
        - HOURS
        - DAYS
      type: string
    TimerScheduleDef:
      description: Schedule definition to execute a job on a timer based configuration
      type: object
      allOf:
        - $ref: '#/components/schemas/ScheduleDef'
      properties:
        startTime:
          type: string
        repeatCount:
          format: int32
          type: integer
        delay:
          format: int64
          type: integer
        delayUnit:
          $ref: '#/components/schemas/TemporalUnitDef'